backend/requirements.txt
fastapi[standard]
pymongo
nanoid
pyjwt